requirements.txt
selenium
requests
openai
python-dotenv
django
djangorestframework